+++ pikpak-plus-client/src/helpers/helpers.ts
@@ -1,8 +1,5 @@
 import { ColDef, SortDirection } from 'ag-grid-community'
-// import { TorrentInfo } from '../types/sharedTypes' // Update the path accordingly
 import axios from 'axios'
-import { logoGithub, heart, link, rocket } from 'ionicons/icons'
-import telegram from '../assets/telegram.svg'
 import { TorrentInfo } from '../types/sharedTypes'
 
 const convertToBytes = (value: string): number => {
@@ -190,47 +187,3 @@ export const formatCreationTime = (creationTime) => {
 
   return formattedDate
 }
-
-export const usefullLinks = [
-  {
-    link: 'Support US',
-    value: 'https://www.buymeacoffee.com/bharathganji',
-    icon: heart,
-  },
-  {
-    link: 'PikPak Invitation (premium)',
-    value:
-      'https://mypikpak.com/drive/activity/invited?invitation-code=47295398',
-    icon: rocket,
-  },
-  {
-    link: 'Github',
-    value: 'https://github.com/bharathganji/pikpak-plus',
-    icon: logoGithub,
-  },
-
-  {
-    link: 'Telegram',
-    value: 'https://t.me/pikpak_plus',
-    icon: telegram,
-  },
-
-  {
-    link: 'Navi Downloader Android',
-    value:
-      'https://github.com/TachibanaGeneralLaboratories/download-navi/releases/',
-    icon: link,
-  },
-  {
-    link: 'VLC Media Player',
-    value: 'https://www.videolan.org/vlc/',
-    icon: link,
-  },
-]
-export const help = [
-  'Download torrent links to Cloud ⚡',
-  'Cumulative download quota 4TB/month',
-  'Storage capacity of 10TB',
-  'Search multiple Torrent Indexers',
-  'Share files with your friends',
-]
